Adjusting Dall-E 3 image generation conditions and flow
Now, the Dall-E 3 function will only be available for users with an 'ADMIN' role or higher. The image generation flow has been improved as well, by downloading and sending the image directly instead of sending a URL. Also, an important condition was added to the settings to prohibit image generation unless specifically asked by the user.